requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
python-dotenv==1.0.0
supabase==2.0.0
requests==2.31.0
PyJWT==2.8.0
bcrypt==4.0.1
python-mailutil==0.0.1
Werkzeug==2.3.6
gunicorn==21.2.0
